+++ src/app/CloudinaryPage/page.tsx
@@ -1,0 +1,10 @@
+import Cloudinary from '@/components/specific/Cloudinary/Cloudinary'
+import React from 'react'
+
+const CloudinaryPage = () => {
+  return (
+     <Cloudinary />
+  )
+}
+
+export default CloudinaryPage
