--- src/app/employee/page.tsx
@@ -1,10 +1,0 @@
-import GetAllUsers from '@/components/specific/Employee/GetAllUsers/GetAllUsers'
-import React from 'react'
-
-function GetAllUsersPage() {
-  return (
-    <GetAllUsers />
-  )
-}
-
-export default GetAllUsersPage
