+++ src/components/specific/Dashboard/UserDashboard.tsx
@@ -4,8 +4,8 @@ import React from "react";
 import { useUserContext } from "@/context/UserContext";
 import DashboardSidebar from "@/components/header/Header";
 import { useUser } from "@auth0/nextjs-auth0/client";
-import CloudinaryPage from "@/app/Cloudinary/page";
 import ProfilePage from "../userData/UserData";
+import Cloudinary from "../Cloudinary/Cloudinary";
 
 const UserDashboard = () => {
   const { userNormal, logoutUser } = useUserContext();
@@ -51,7 +51,7 @@ const UserDashboard = () => {
                 Cerrar sesión
               </button>
             </div>
-            <CloudinaryPage userprops={{ email: userNormal?.email, image_url: userNormal?.image_url }} />
+            <Cloudinary />
           </div>
         ) : (
           <div className="text-center mt-20">
